Download template R scripts

This function will download R script templates for tasks commonly used in the EngleLab

get_script(type = "all", to = "R Scripts", overwrite = FALSE,
  gf = FALSE, wmc = FALSE, ac = FALSE, antisaccade = FALSE,
  stroop = FALSE, flanker = FALSE, stroopDL = FALSE,
  flankerDL = FALSE, va4 = FALSE, sact = FALSE, path = "./")

Arguments

type
String. Do you want to download "raw" or "score" task template files?

to
a directory where to download R scripts

gf
Logical. Do you want to download template scripts for the gf taks?

wmc
Logical. Do you want to download template scripts for the wmc taks?

ac
Logical. Do you want to download template scripts for all the attention taks?

antisaccade
Logical. Do you want to download a template file for the antisaccade task?

stroop
Logical. Do you want to download a template file for the Stroop task?

flanker
Logical. Do you want to download a template file for the flanker task?

stroopDL
Logical. Do you want to download a template file for the StroopDL task?

flankerDL
Logical. Do you want to download a template file for the FlankerDL task?

va4
Logical. Do you want to download a template file for the Visual Arrays 4 task?

sact
Logical. Do you want to download a template file for the SACT task?

path
String. Home directory file path